+++ ding/entry/main.py
@@ -2,24 +2,23 @@
 Main entry
 """
 from collections import deque
-from types import GeneratorType
-import gym
 import torch
 import numpy as np
 import time
 from rich import print
 from functools import partial
-from ding.model import DQN
+from ding.model import QAC, DQN
 from ding.utils import set_pkg_seed
-from ding.envs import DingEnvWrapper, BaseEnvManager, get_vec_env_setting
+from ding.envs import DingEnvWrapper, BaseEnvManager, get_vec_env_setting, SyncSubprocessEnvManager
 from ding.config import compile_config
-from ding.policy import DQNPolicy
-from ding.rl_utils import get_epsilon_greedy_fn
+from ding.policy import SACPolicy, DQNPolicy
 from ding.torch_utils import to_ndarray, to_tensor
+from ding.rl_utils import get_epsilon_greedy_fn
 from ding.worker.collector.base_serial_evaluator import VectorEvalMonitor
-from ding.framework import Task
-from dizoo.classic_control.cartpole.config.cartpole_dqn_config import main_config, create_config
-# from dizoo.atari.config.serial.pong.pong_dqn_config import main_config, create_config
+from ding.framework import Task, Parallel
+from ding.framework.wrapper import StepTimer
+# from dizoo.classic_control.pendulum.config.pendulum_sac_config import main_config, create_config
+from dizoo.atari.config.serial.pong.pong_dqn_config import main_config, create_config
 
 
 class DequeBuffer:
@@ -47,36 +46,87 @@ class DequeBuffer:
         return len(self.memory)
 
 
-class DQNPipeline:
+def differential(task, model, buffer: DequeBuffer, wait_n_sample: int, collect_before_wait: int):
+    """
+    Sync train/collect/evaluate speed
+    """
+
+    cum_buffer_count = 0
+    receive_new_model = False
+    collected_before_wait = 0
+    finish = False
+
+    def sync_parallel_ctx(ctx):
+        nonlocal cum_buffer_count, receive_new_model, finish
+        if ctx.finish:
+            finish = ctx.finish
+        if "collect_transitions" in ctx:
+            for transition in ctx.collect_transitions:
+                buffer.push(transition)
+                cum_buffer_count += 1
+        if "model_weight" in ctx:
+            model.load_state_dict(ctx.model_weight)
+            receive_new_model = True
+
+    task.on("sync_parallel_ctx", sync_parallel_ctx)
+
+    def _differential(ctx):
+        print("Model", task.router.node_id, model.state_dict()['head.V.1.0.weight'][0][:10])
+        nonlocal cum_buffer_count, receive_new_model, collected_before_wait, finish
+        timeout = 5
+        if finish:
+            ctx.finish = finish
+        if "train_iter" in ctx:  # On learner
+            # Wait until buffer has enough data
+            for _ in range(timeout * 100):
+                if cum_buffer_count < wait_n_sample:
+                    time.sleep(0.01)
+                else:
+                    break
+            cum_buffer_count = 0
+        elif "collect_env_step" in ctx:  # On collector
+            collected_before_wait += len(ctx.get("collect_transitions") or [])
+            # Wait until receive new model
+            for _ in range(timeout * 100):
+                if collected_before_wait > collect_before_wait and not receive_new_model:
+                    time.sleep(0.01)
+                else:
+                    break
+            if collected_before_wait > collect_before_wait:
+                collected_before_wait = 0
+            if receive_new_model:
+                receive_new_model = False
+
+    return _differential
+
+
+class Pipeline:
 
     def __init__(self, cfg, model: torch.nn.Module):
         self.cfg = cfg
         self.model = model
+        # self.policy = SACPolicy(cfg.policy, model=model)
         self.policy = DQNPolicy(cfg.policy, model=model)
-        eps_cfg = cfg.policy.other.eps
-        self.epsilon_greedy = get_epsilon_greedy_fn(eps_cfg.start, eps_cfg.end, eps_cfg.decay, eps_cfg.type)
+        if 'eps' in cfg.policy.other:
+            eps_cfg = cfg.policy.other.eps
+            self.epsilon_greedy = get_epsilon_greedy_fn(eps_cfg.start, eps_cfg.end, eps_cfg.decay, eps_cfg.type)
 
     def act(self, env):
 
         def _act(ctx):
             ctx.setdefault("collect_env_step", 0)
             ctx.keep("collect_env_step")
-            eps = self.epsilon_greedy(ctx.collect_env_step)
             ctx.obs = env.ready_obs
-            policy_output = self.policy.collect_mode.forward(ctx.obs, eps=eps)
+            policy_kwargs = {}
+            if hasattr(self, 'epsilon_greedy'):
+                policy_kwargs['eps'] = self.epsilon_greedy(ctx.collect_env_step)
+            policy_output = self.policy.collect_mode.forward(ctx.obs, **policy_kwargs)
             ctx.action = to_ndarray({env_id: output['action'] for env_id, output in policy_output.items()})
             ctx.policy_output = policy_output
 
         return _act
 
     def collect(self, env, buffer_, task: Task):
-
-        def on_sync_parallel_ctx(ctx):
-            if "collect_transitions" in ctx:
-                for t in ctx.collect_transitions:
-                    buffer_.push(t)
-
-        # task.on("sync_parallel_ctx", on_sync_parallel_ctx)
 
         def _collect(ctx):
             timesteps = env.step(ctx.action)
@@ -92,7 +142,7 @@ class DQNPipeline:
 
         return _collect
 
-    def learn(self, buffer_, task: Task):
+    def learn(self, buffer_: DequeBuffer, task: Task):
 
         def _learn(ctx):
             ctx.setdefault("train_iter", 0)
@@ -147,15 +197,16 @@ class DQNPipeline:
         return _eval
 
 
-def sample_profiler(buffer, print_per_step=1, async_mode=False):
+def sample_profiler(buffer, print_per_step=1):
     start_time = None
     start_counter = 0
     start_step = 0
-    records = deque(maxlen=10)
-    step_records = deque(maxlen=10)
+    records = deque(maxlen=print_per_step * 50)
+    step_records = deque(maxlen=print_per_step * 50)
+    max_mean = 0
 
     def _sample_profiler(ctx):
-        nonlocal start_time, start_counter, start_step
+        nonlocal start_time, start_counter, start_step, max_mean
         if not start_time:
             start_time = time.time()
         elif ctx.total_step % print_per_step == 0:
@@ -166,146 +217,93 @@ def sample_profiler(buffer, print_per_step=1, async_mode=False):
             step_record = (end_step - start_step) / (end_time - start_time)
             records.append(record)
             step_records.append(step_record)
+            max_mean = max(np.mean(records), max_mean)
             print(
-                "        Samples/s: {:.2f}, Mean: {:.2f}, Total: {:.0f}; Steps/s: {:.2f}, Mean: {:.2f}, Total: {:.0f}".
-                format(record, np.mean(records), end_counter, step_record, np.mean(step_records), ctx.total_step)
+                "        Samples/s: {:.2f}, Mean: {:.2f}, Max: {:.2f}, Total: {:.0f};\
+ Steps/s: {:.2f}, Mean: {:.2f}, Total: {:.0f}".format(
+                    record, np.mean(records), max_mean, end_counter, step_record, np.mean(step_records), ctx.total_step
+                )
             )
             start_time, start_counter, start_step = end_time, end_counter, end_step
 
-    async def async_sample_profiler(ctx):
-        _sample_profiler(ctx)
-
-    return async_sample_profiler if async_mode else _sample_profiler
+    return _sample_profiler
 
 
-def step_profiler(step_name, silent=False):
-    records = deque(maxlen=10)
+class Reporter():
 
-    def _step_wrapper(fn):
-        # Wrap step function
-        def _step_executor(ctx):
-            # Execute step
-            start_time = time.time()
-            time_cost = 0
-            g = fn(ctx)
-            if isinstance(g, GeneratorType):
-                next(g)
-                time_cost = time.time() - start_time
-                yield
-                start_time = time.time()
-                try:
-                    next(g)
-                except StopIteration:
-                    pass
-                time_cost += time.time() - start_time
-            else:
-                time_cost = time.time() - start_time
-            records.append(time_cost * 1000)
-            if not silent:
-                print(
-                    "    Step Profiler {}: Cost: {:.2f}ms, Mean: {:.2f}ms".format(
-                        step_name, time_cost * 1000, np.mean(records)
-                    )
-                )
+    def __init__(self, task_name, async_mode, parallel_mode, seed):
+        # Calculate execution time
+        # Save statistics to report.txt
+        self.start = time.time()
+        self.task_name = task_name
+        self.async_mode = async_mode
+        self.parallel_mode = parallel_mode
+        self.seed = seed
+        self.train_iter = 0
+        self.collect_env_step = 0
 
-        return _step_executor
-
-    return _step_wrapper
+    def record(self):
+        duration = time.time() - self.start
+        template = "task:{},seed:{},async:{},parallel:{},train_iter:{},env_step:{},duration:{:.2f}"
+        with open("./tmp/report.txt", "a+") as f:
+            report = template.format(
+                self.task_name, self.seed, self.async_mode, self.parallel_mode, self.train_iter, self.collect_env_step,
+                self.duration
+            )
+            print(report)
+            f.write(report + "\n")
 
 
-def mock_pipeline(buffer):
+def main(cfg, model, async_mode, seed=0):
+    with Task(async_mode=async_mode) as task:
+        env_fn, collector_env_cfg, evaluator_env_cfg = get_vec_env_setting(cfg.env)
 
-    def _mock_pipeline(ctx):
-        buffer.push(0)
+        collector_env = BaseEnvManager(env_fn=[partial(env_fn, cfg=c) for c in collector_env_cfg], cfg=cfg.env.manager)
+        evaluator_env = BaseEnvManager(env_fn=[partial(env_fn, cfg=c) for c in evaluator_env_cfg], cfg=cfg.env.manager)
 
-    return _mock_pipeline
+        collector_env.seed(seed)
+        evaluator_env.seed(seed, dynamic_seed=False)
+        set_pkg_seed(seed, use_cuda=cfg.policy.cuda)
+        collector_env.launch()
+        evaluator_env.launch()
 
+        replay_buffer = DequeBuffer()
+        sac = Pipeline(cfg, model)
 
-def main(cfg, create_cfg, seed=0):
-
-    def wrapped_cartpole_env():
-        return DingEnvWrapper(gym.make('CartPole-v0'))
-
-    cfg = compile_config(cfg, create_cfg=create_cfg, auto=True)
-    collector_env_num, evaluator_env_num = cfg.env.collector_env_num, cfg.env.evaluator_env_num
-    collector_env = BaseEnvManager(env_fn=[wrapped_cartpole_env for _ in range(collector_env_num)], cfg=cfg.env.manager)
-    evaluator_env = BaseEnvManager(env_fn=[wrapped_cartpole_env for _ in range(evaluator_env_num)], cfg=cfg.env.manager)
-
-    collector_env.seed(seed)
-    evaluator_env.seed(seed, dynamic_seed=False)
-    set_pkg_seed(seed, use_cuda=cfg.policy.cuda)
-    collector_env.launch()
-    evaluator_env.launch()
-
-    model = DQN(**cfg.policy.model)
-    replay_buffer = DequeBuffer()
-
-    task = Task()
-    dqn = DQNPipeline(cfg, model)
-
-    task.use(sample_profiler(replay_buffer))
-    task.use(dqn.evaluate(evaluator_env))
-    task.use(task.sequence(dqn.act(collector_env), dqn.collect(collector_env, replay_buffer)))
-    task.use(dqn.learn(replay_buffer))
-
-    task.run(max_step=1000)
-
-
-def print_step(task: Task):
-    import random
-    from os import path
-    time.sleep(random.random() + 1)
-    print(
-        "Current task step on {}".format(task.parallel_mode and path.basename(task.router._bind_addr or "")),
-        task.ctx.total_step
-    )
-
-
-def main_eager(cfg, create_cfg, seed=0):
-
-    cfg = compile_config(cfg, create_cfg=create_cfg, auto=True)
-    task = Task(async_mode=True, n_async_workers=3, parallel_mode=True, n_parallel_workers=2, attach_to=[])
-    start = time.time()
-
-    env_fn, collector_env_cfg, evaluator_env_cfg = get_vec_env_setting(cfg.env)
-
-    collector_env = BaseEnvManager(env_fn=[partial(env_fn, cfg=c) for c in collector_env_cfg], cfg=cfg.env.manager)
-    evaluator_env = BaseEnvManager(env_fn=[partial(env_fn, cfg=c) for c in evaluator_env_cfg], cfg=cfg.env.manager)
-
-    collector_env.seed(seed)
-    evaluator_env.seed(seed, dynamic_seed=False)
-    set_pkg_seed(seed, use_cuda=cfg.policy.cuda)
-    collector_env.launch()
-    evaluator_env.launch()
-
-    model = DQN(**cfg.policy.model)
-    # model.share_memory()
-    replay_buffer = DequeBuffer()
-
-    dqn = DQNPipeline(cfg, model)
-    evaluate = dqn.evaluate(evaluator_env)
-    act = dqn.act(collector_env)
-    collect = dqn.collect(collector_env, replay_buffer, task=task)
-    learn = dqn.learn(replay_buffer, task=task)
-    profiler = sample_profiler(replay_buffer, print_per_step=50)
-
-    def _execute_task(task: Task):
-        while task.ctx.total_step < 1000:
-            task.forward(profiler)
-            task.forward(evaluate)
-            if task.finish:
-                break
-            task.forward(task.sequence(act, collect))
-            task.forward(learn)
-            task.renew()
-            # print_step(task)
-        # print("model weight", model._version, model.state_dict()["encoder.main.0.weight"][0][:10])
-
-    task.parallel(_execute_task)
-    # _execute_task(task)
-    print("Total time cost: {:.2f}s".format(time.time() - start))
+        task.use_step_wrapper(StepTimer(print_per_step=1))
+        task.use(sac.evaluate(evaluator_env), filter_labels=["standalone", "node.0"])
+        task.use(
+            task.sequence(sac.act(collector_env), sac.collect(collector_env, replay_buffer, task=task)),
+            filter_labels=["standalone", "node.[1-9]*"]
+        )
+        task.use(sac.learn(replay_buffer, task=task), filter_labels=["standalone", "node.0"])
+        task.use(
+            differential(task, model, replay_buffer, wait_n_sample=96, collect_before_wait=48),
+            filter_labels=["distributed"]
+        )
+        task.run(max_step=10000)
+        r.train_iter = task.ctx.train_iter
+        r.collect_env_step = task.ctx.collect_env_step
 
 
 if __name__ == "__main__":
-    # main(main_config, create_config)
-    main_eager(main_config, create_config)
+    from ding.utils import Profiler
+    Profiler().profile()
+
+    import os
+    task_name = "Pong/DQN"
+    async_mode = False
+    parallel_mode = False
+    seed = int(os.environ.get("SEED")) if os.environ.get("SEED") else 0
+
+    r = Reporter(task_name, async_mode, parallel_mode, seed)
+    cfg = compile_config(main_config, create_cfg=create_config, auto=True)
+    model = DQN(**cfg.policy.model)
+
+    if not parallel_mode:
+        main(cfg, model, async_mode, seed)
+    else:
+        n_parallel_workers = 3
+        cfg["env"]["collector_env_num"] //= n_parallel_workers - 1
+        Parallel.runner(n_parallel_workers=n_parallel_workers, topology="star")(main, cfg, model, async_mode, seed)
+    r.record()
